Add files via upload
This is the R code used to run the linear mixed effects analyses on my paper "Multimodal parent behaviors within Joint Attention support Sustained Attention in infants". This code uses the data called "lme_data.csv" to run the analyses although the data only includes 100 rows and is thus meant to be a sample data for this type of analyses.